Misc Fixes/Features
The date error is fixed. If the chapsPerDay is less than 1 it will
reduce the number of days s the minimum chapters is always 1.  When the
user completes the last reading assignment it saves a variable named
finished to the schedule and if the main activity finds a finished
schedule it will delete the file. Added a lastmodified variable.